Adding the Failer Class
The class is responsible for maintaining the failer list through Failerlist.txt file in the project folder. The file will contain the information of tickers which were interupted while processing due to the issues in CNN, Yahoo sites. 
If the program finds that the data regarding a ticker is unavailable, the ticker will be added to the Failerlist.txt file. Once the program finishes the Ticker sorting part, it checks if the status for any of the ticker has changed and accordingly updates the Failerlist.txt file.